feat: improve output when polling URLs
